mv repeat_kv and repeat_kv_bshd to common
